py_backend/requirements.txt
fastapi
uvicorn
python-multipart
deepface
Pillow
